update doc and use mock generator for tests
