+++ hi-ml-cpath/testhisto/testhisto/models/test_deepmil.py
@@ -134,27 +134,6 @@ def _test_lightningmodule(
 
 
 @pytest.fixture(scope="session")
-def mock_panda_tiles_root_dir(
-    tmp_path_factory: pytest.TempPathFactory, tmp_path_to_pathmnist_dataset: Path
-) -> Generator:
-    tmp_root_dir = tmp_path_factory.mktemp("mock_tiles")
-    tiles_generator = MockPandaTilesGenerator(
-        dest_data_path=tmp_root_dir,
-        src_data_path=tmp_path_to_pathmnist_dataset,
-        mock_type=MockHistoDataType.PATHMNIST,
-        n_tiles=4,
-        n_slides=10,
-        n_channels=3,
-        tile_size=28,
-        img_size=224,
-    )
-    logging.info("Generating temporary mock tiles that will be deleted at the end of the session.")
-    tiles_generator.generate_mock_histo_data()
-    yield tmp_root_dir
-    shutil.rmtree(tmp_root_dir)
-
-
-@pytest.fixture(scope="session")
 def mock_panda_slides_root_dir(
     tmp_path_factory: pytest.TempPathFactory, tmp_path_to_pathmnist_dataset: Path
 ) -> Generator:
